moving unit tests to integration tests and remove duplicates
